// packages/server/src/proto/compiled.d.ts
export namespace transit_realtime {
// ...
  interface IFeedHeader {
    /** FeedHeader gtfsRealtimeVersion */
    gtfsRealtimeVersion: string;

    /** FeedHeader incrementality */
    incrementality?: transit_realtime.FeedHeader.Incrementality | null;

    /** FeedHeader timestamp */
    timestamp?: number | Long | null;

    /** FeedHeader .transit_realtime.nyctFeedHeader */
    ".transit_realtime.nyctFeedHeader"?: transit_realtime.INyctFeedHeader | null;
  }
// ...
  namespace TripUpdate {
    /** Properties of a StopTimeEvent. */
    interface IStopTimeEvent {
      /** StopTimeEvent delay */
      delay?: number | null;

      /** StopTimeEvent time */
      time?: number | Long | null;

      /** StopTimeEvent uncertainty */
      uncertainty?: number | null;
    }
// ...
    interface IStopTimeUpdate {
      /** StopTimeUpdate stopSequence */
      stopSequence?: number | null;

      /** StopTimeUpdate stopId */
      stopId?: string | null;

      /** StopTimeUpdate arrival */
      arrival?: transit_realtime.TripUpdate.IStopTimeEvent | null;

      /** StopTimeUpdate departure */
      departure?: transit_realtime.TripUpdate.IStopTimeEvent | null;

      /** StopTimeUpdate scheduleRelationship */
      scheduleRelationship?: transit_realtime.TripUpdate.StopTimeUpdate.ScheduleRelationship | null;

      /** StopTimeUpdate stopTimeProperties */
      stopTimeProperties?: transit_realtime.TripUpdate.StopTimeUpdate.IStopTimeProperties | null;

      /** StopTimeUpdate .transit_realtime.nyctStopTimeUpdate */
      ".transit_realtime.nyctStopTimeUpdate"?: transit_realtime.INyctStopTimeUpdate | null;
    }
// ...
    namespace StopTimeUpdate {
      /** ScheduleRelationship enum. */
      enum ScheduleRelationship {
        SCHEDULED = 0,
        SKIPPED = 1,
        NO_DATA = 2,
        UNSCHEDULED = 3,
      }

      /** Properties of a StopTimeProperties. */
      interface IStopTimeProperties {
        /** StopTimeProperties assignedStopId */
        assignedStopId?: string | null;
      }
// ...
    }
// ...
  }
// ...
  interface ITripReplacementPeriod {
    /** TripReplacementPeriod routeId */
    routeId?: string | null;

    /** TripReplacementPeriod replacementPeriod */
    replacementPeriod?: transit_realtime.ITimeRange | null;
  }
// ...
  interface INyctFeedHeader {
    /** NyctFeedHeader nyctSubwayVersion */
    nyctSubwayVersion: string;

    /** NyctFeedHeader tripReplacementPeriod */
    tripReplacementPeriod?: transit_realtime.ITripReplacementPeriod[] | null;
  }
// ...
  namespace NyctTripDescriptor {
    /** Direction enum. */
    enum Direction {
      NORTH = 1,
      EAST = 2,
      SOUTH = 3,
      WEST = 4,
    }
  }

  /** Properties of a NyctStopTimeUpdate. */
  interface INyctStopTimeUpdate {
    /** NyctStopTimeUpdate scheduledTrack */
    scheduledTrack?: string | null;

    /** NyctStopTimeUpdate actualTrack */
    actualTrack?: string | null;
  }
// ...
}